@@ -20,13 +20,12 @@ setup(
     use_scm_version={"local_scheme": local_scheme},
     setup_requires=["setuptools_scm"],
     extras_require={"tests": ["coverage", "pillow", "pytest", "pytest-cov"]},
-    python_requires=">=3.5",
+    python_requires=">=3.6",
     classifiers=[
         # 'Development Status :: 5 - Production/Stable',
         "Intended Audience :: Developers",
         "License :: OSI Approved :: MIT License",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
